+++ src/kerko/views/__init__.py
@@ -1,3 +1,1 @@
-"""Routed views."""
-
-from kerko.views import routes  # noqa: F401
+"""Views."""