+++ tests/test_pager.py
@@ -18,7 +18,7 @@ class SectionsTestCase(unittest.TestCase):
     def setUp(self):
         self.app = Flask(__name__)
         config_update(self.app.config, kerko.DEFAULTS)
-        self.app.register_blueprint(kerko.blueprint, url_prefix='/bibliography')
+        self.app.register_blueprint(kerko.make_blueprint(), url_prefix='/bibliography')
         ctx = self.app.app_context()
         ctx.push()